setup the general workflow of the application
There's a lot of work in here, but don't worry, we'll be going with smaller commits from here on out. This commit really just setups up the ideal workflow.

What we want to do is generate a payout when a Github merge occurs. To do this, we need to accept a webhook, make sure it's a merge event from a PR, log it as a Merge, and then execute a payout job.

For payouts, we want to check if the user has supplied their bitcoin address. We don't want to just accept it from Github and put them on the spot. We'll ask for it privately via email. Then we should ideally rerun the payout job.

If there is a bitcoin address, we're going to want to calculate an amount. We can do this either by checking the user's toggle or by checking if they solved a bounty. We need logic for this, but for not we have the pseudocode. If everything goes to plan, we should be able to create a payout for them and then send a notification to the admins.

The admins will need to reply back (ideally on Slack) and say that they approve it. We should create an approvals controller and connect that to Slackbot. Once the approval has been made, boom! We can send money out to the user.

For sending bitcoin, I'm not sure what platform we're going to use yet.